Explicitly cast the return value
When this library was used with bios-bmc-smm-error-logger, it was
failing because of a clang-tidy warning(For some reason clang-tidy
runs on subprojects/libbej as well.).

We need to explicitly cast to stop this warning.

Tested:
Tested running CI locally for bios-bmc-smm-error-logger with this
change and it passed.

+++ src/bej_decoder_core.c
@@ -62,7 +62,7 @@ static int64_t bejGetIntegerValue(const uint8_t* bytes, uint8_t numOfBytes)
     uint8_t bitsInVal = numOfBytes * 8;
     // Since numOfBytes > 0, bitsInVal is non negative.
     uint64_t mask = (uint64_t)1 << (uint8_t)(bitsInVal - 1);
-    return (value ^ mask) - mask;
+    return (int64_t)((value ^ mask) - mask);
 }
 
 /**
